requirements.txt
aiogram==2.19
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
dnspython==2.2.1
email-validator==1.1.3
environs==9.5.0
Faker==13.3.2
fastapi==0.75.0
fastapi-users==9.3.0
fastapi-users-db-sqlalchemy==3.0.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.0.3
makefun==1.13.1
MarkupSafe==2.1.1
marshmallow==3.15.0
multidict==6.0.2
packaging==21.3
passlib==1.7.4
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.6
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8
websockets==10.2
WTForms==3.0.1
yarl==1.7.2
